update: strings res & surah activity

+++ app/src/main/java/com/codetarian/bacaquran/activity/SurahActivity.kt
@@ -3,18 +3,18 @@ package com.codetarian.bacaquran.activity
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
 import com.codetarian.bacaquran.adapter.ViewPagerAdapter
-import com.codetarian.bacaquran.databinding.ActivityMainBinding
+import com.codetarian.bacaquran.databinding.ActivitySurahBinding
 import com.codetarian.bacaquran.fragment.SurahFragment
 import java.io.FileOutputStream
 import java.io.IOException
 
-class MainActivity : AppCompatActivity() {
+class SurahActivity : AppCompatActivity() {
 
-    private lateinit var binding: ActivityMainBinding
+    private lateinit var binding: ActivitySurahBinding
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         copyDatabaseFromAssets()
-        binding = ActivityMainBinding.inflate(layoutInflater)
+        binding = ActivitySurahBinding.inflate(layoutInflater)
         setContentView(binding.root)
 
         binding.apply {
@@ -29,16 +29,16 @@ class MainActivity : AppCompatActivity() {
 
     private fun copyDatabaseFromAssets() {
         val databaseFile = getDatabasePath("quran_database.db")
-//        if (!databaseFile.exists()) {
-        try {
-            val inputStream = assets.open("quran_database.db")
-            val outputStream = FileOutputStream(databaseFile)
-            inputStream.copyTo(outputStream)
-            inputStream.close()
-            outputStream.close()
-        } catch (e: IOException) {
-            e.printStackTrace()
+        if (!databaseFile.exists()) {
+            try {
+                val inputStream = assets.open("quran_database.db")
+                val outputStream = FileOutputStream(databaseFile)
+                inputStream.copyTo(outputStream)
+                inputStream.close()
+                outputStream.close()
+            } catch (e: IOException) {
+                e.printStackTrace()
+            }
         }
-//        }
     }
 }
